+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 
 #### Added
 - Added TestStdIO.
+- Added tests for Release build.
 - Added shrinked [diagnostics](https://github.com/dotnet/diagnostics) sources v9.0.661903, dbgshim library build now during debugger build.
 - Added clang-tidy checks.
 - Added cppcheck checks.